apps/payment/views/create_order.py
from rest_framework.generics import CreateAPIView
from rest_framework.permissions import IsAuthenticated
from django.db import transaction

from apps.payment.models import Order, Transaction, Provider
from apps.payment.choices import OrderStatus, TransactionStatus
from apps.payment.serializers import OrderCreateSerializer


class OrderCreateAPIView(CreateAPIView):
    queryset = Order.objects.all()
    serializer_class = OrderCreateSerializer
    permission_classes = [IsAuthenticated]