pyproject.toml
[project]
name = "django-boilerplate-modern"
version = "0.1.0"
description = "Modern Django Boilerplate"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "daphne>=4.2.1",
    "django>=5.2.3",
    "django-jazzmin>=3.0.1",
    "django-recaptcha>=4.1.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.10",
    "flower>=2.0.1",
    "httpx>=0.28.1",
    "pillow>=11.3.0",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "ruff>=0.12.0",
]


# RUFF configs
[tool.ruff]
# General settings
target-version = "py312"
line-length = 88

# Enable rules
lint.select = ["E", "F", "I", "N", "UP", "B", "C90", "RUF"]

lint.ignore = [
    "E501",     # line too long
    "RUF001",   # ambiguous unicode character in code
    "RUF002",   # ambiguous unicode character in docstring
]  

# Paths to exclude
exclude = ["migrations", "__pycache__", ".venv"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra --tb=short --strict-markers"
testpaths = ["tests"]
